create two separate route mixins
