+++ static/js/src/entry/user-portal/mixins/route-fetch.js
@@ -1,0 +1,19 @@
+import contextMixin from './context';
+
+export default {
+  mixins: [contextMixin],
+
+  beforeMount() {
+    this.context.setIsFetching(true);
+  },
+
+  async mounted() {
+    try {
+      await this.fetchData();
+    } catch (err) {
+      this.context.setError(true);
+    }
+
+    this.context.setIsFetching(false);
+  },
+};
