+++ static/js/src/entry/user-portal/routes/blast/Index.vue
@@ -4,11 +4,12 @@
 
 <script>
 import routeMixin from '../../mixins/route';
+import routeFetchMixin from '../../mixins/route-fetch';
 
 export default {
   name: 'Index',
 
-  mixins: [routeMixin],
+  mixins: [routeMixin, routeFetchMixin],
 
   methods: {
     async fetchData() {
